+++ shared/model/constants.go
@@ -1,0 +1,11 @@
+package model
+
+const (
+	ContentTypeMovie   = "movie"
+	ContentTypeEpisode = "episode"
+
+	SourceTypeTorrent = "torrent"
+	SourceTypeUpload  = "upload"
+	SourceTypeHTTP    = "http"
+	SourceTypeIngest  = "ingest"
+)